refactor: use module-specific fixtures for tests
