+++ test/app.e2e-spec.ts
@@ -2,11 +2,8 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication, ValidationPipe } from '@nestjs/common';
 import * as request from 'supertest';
 import { AppModule } from './../src/app.module';
-import { GlobalFixtures } from './fixtures/global-fixtures';
-
 describe('AppController (e2e)', () => {
   let app: INestApplication;
-  let fixtures: GlobalFixtures;
 
   beforeAll(async () => {
     const moduleFixture: TestingModule = await Test.createTestingModule({
@@ -23,14 +20,9 @@ describe('AppController (e2e)', () => {
     );
     app.setGlobalPrefix('api');
     await app.init();
-
-    // Initialize fixtures
-    fixtures = new GlobalFixtures(app);
-    await fixtures.load();
   });
 
   afterAll(async () => {
-    await fixtures.clear();
     await app.close();
   });
